# utme-master-backend/docs/api-spec.yaml
...
openapi: 3.0.3
info:
  title: UTME Master API
  description: |
    UTME Master Backend API - Exam preparation system for JAMB, WAEC, NECO
    
    ## Authentication
    All protected endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```
    
    ## Response Format
    All responses follow a consistent format:
    ```json
    {
      "success": true,
      "data": { ... },
      "message": "Operation successful"
    }
    ```
    
    ## Error Responses
    Common error codes:
    - `400`: Bad Request - Invalid input
    - `401`: Unauthorized - Missing or invalid token
    - `403`: Forbidden - Insufficient permissions
    - `404`: Not Found - Resource doesn't exist
    - `500`: Internal Server Error
    
  version: 1.0.0
  contact:
    name: UTME Master Support
    email: support@utmemaster.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.utmemaster.com/api
    description: Production server
...
paths:
...
  /questions:
    get:
      tags:
        - Questions
      summary: Get questions
      description: Retrieve questions with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: subject
          in: query
          schema:
            type: string
            example: "Mathematics"
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [EASY, MEDIUM, HARD]
            example: "EASY"
        - name: examType
          in: query
          schema:
            type: string
            enum: [JAMB, WAEC, NECO, MOCK, PRACTICE]
            example: "JAMB"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Questions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      questions:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              example: "question-123"
                            questionText:
                              type: string
                              example: "What is the capital of Nigeria?"
                            questionType:
                              type: string
                              example: "MCQ"
                            difficulty:
                              type: string
                              example: "EASY"
                            subject:
                              type: string
                              example: "Mathematics"
                            isActive:
                              type: boolean
                              example: true
                      total:
                        type: integer
                        example: 100
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 20
  /questions/{questionId}:
    get:
      tags:
        - Questions
      summary: Get question details
      description: Retrieve detailed information about a specific question
      security:
        - bearerAuth: []
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
            example: "question-123"
      responses:
        '200':
          description: Question details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "question-123"
                      questionText:
                        type: string
                        example: "What is the capital of Nigeria?"
                      questionType:
                        type: string
                        example: "MCQ"
                      options:
                        type: object
                        properties:
                          A:
                            type: object
                            properties:
                              text:
                                type: string
                                example: "Lagos"
                          B:
                            type: object
                            properties:
                              text:
                                type: string
                                example: "Abuja"
                          C:
                            type: object
                            properties:
                              text:
                                type: string
                                example: "Kano"
                          D:
                            type: object
                            properties:
                              text:
                                type: string
                                example: "Port Harcourt"
                      correctAnswer:
                        type: string
                        example: "B"
                      explanation:
                        type: string
                        example: "Abuja is the capital of Nigeria"
                      difficulty:
                        type: string
                        example: "EASY"
                      subject:
                        type: string
                        example: "Mathematics"
  
  # ==========================================
  # RESULTS
  # ==========================================
...